+++ package.json
@@ -18,6 +18,7 @@
     "@types/react-dom": "^17.0.0",
     "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.7",
+    "axios": "^0.21.4",
     "dayjs": "^1.10.4",
     "firebase": "^8.3.2",
     "formik": "^2.2.6",
