+++ src/utils/firebase/firebase-init.ts
@@ -12,5 +12,6 @@ const firebaseConfig = {
 };
 
 firebase.initializeApp(firebaseConfig);
+firebase.firestore().enablePersistence();
 
 export default firebase.firestore();
